requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
bcrypt==4.2.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloud-sql-python-connector==1.16.0
cryptography==44.0.2
dnspython==2.7.0
e==1.4.5
ecdsa==0.19.1
email-validator==2.2.0
et-xmlfile==2.0.0
event-stream @ file:///Users/nayz/Desktop/Taxibee/event_stream
fastapi==0.115.8
frozenlist==1.5.0
google-auth==2.38.0
h11==0.14.0
idna==3.10
kafka-python==2.1.5
multidict==6.2.0
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pip==25.0
propcache==0.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pymysql==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
pytz==2025.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.38
starlette==0.45.3
typing-extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
